# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sera Test API",
        "contact": {
            "name": "Ibnu Muntoha",
            "email": "ibnu.muntoha@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/company/": {
            "post": {
                "tags": [
                    "Company Using Firestore"
                ],
                "summary": "Create company By Id",
                "description": "Creates a new company data",
                "operationId": "Add company",
                "requestBody": {
                    "description": "to add company data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jakarta"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jaya Abadi"
                                    },
                                    "address": {
                                        "properties": {
                                            "building": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "230"
                                            },
                                            "street": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "Jl.Kemerdekaan"
                                            },
                                            "zipcode": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "14502"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Created"
                                        },
                                        "collection_id": {
                                            "type": "string",
                                            "example": "azgZaZT7bCdInSAymsHgyVy0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/company/{id}": {
            "get": {
                "tags": [
                    "Company Using Firestore"
                ],
                "summary": "Get Company By Id",
                "description": "Get Data company By spesicif document id",
                "operationId": "/retaurant/id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id parameter in path is document id in Firestore ex: SURvO6l5ILojZ9i3QeYfn8k0",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns Company data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Get Data"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "some company  data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Company Using Firestore"
                ],
                "summary": "Update company By Id",
                "description": "Update company data",
                "operationId": "Update company",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "to update company data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jakarta"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jaya Abadi"
                                    },
                                    "address": {
                                        "properties": {
                                            "building": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "230"
                                            },
                                            "street": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "Jl.Kemerdekaan"
                                            },
                                            "zipcode": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "14502"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "pet response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Company Using Firestore"
                ],
                "summary": "Delete company By Id",
                "description": "Delete Data company By spesicif document id",
                "operationId": "/retaurant/id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id parameter in path is document id ex: SURvO6l5ILojZ9i3QeYfn8k0",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns company data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}